add "continue" response to pager
This adds a "continue" response to the pager.  If the user types "c"
in response to the pager prompt, pagination will be disabled for the
duration of one command -- but re-enabled afterward.  This is handy if
you type a command that produces a lot of output, and you don't want
to baby-sit it by typing "return" each time the prompt comes up.

Built and regtested on x86-64 Fedora 20.
This required a couple of minor tweaks to existing tests.

2014-06-24  Tom Tromey  <tromey@redhat.com>

	PR cli/12326:
	* utils.c (pagination_disabled_for_command): New global.
	(prompt_for_continue): Allow "c" response to prompt.
	(reinitialize_more_filter): Clear
	pagination_disabled_for_command.
	(fputs_maybe_filtered): Check pagination_disabled_for_command.

2014-06-24  Tom Tromey  <tromey@redhat.com>

	* gdb.texinfo (Screen Size): Document "c" response to pagination
	prompt.

2014-06-24  Tom Tromey  <tromey@redhat.com>

	* gdb.base/page.exp: Update expected output.  Add new tests.
	* gdb.python/python.exp: Update expected output.