Fix PR NNN - add __index__ method to gdb.Value
!!! FIXME !!!
 - needs regtest
 - needs PR #, plus in changelog
 - perhaps also fix hex()

This patch fixes PR NNNN.

The bug is that gdb.Value does not implement the Python __index__
method.  This method is needed to convert a Python object to an index
and is used by various operations in Python, such as indexing an
array.

The fix is to implement the nb_index method for gdb.Value.

While doing this I discovered that value_object_as_number was missing
an entry.  This patch fixes this.

nb_index was added in Python 2.5.  I don't have a good way to test
Python 2.4, but I made an attempt.

I chose to use valpy_long in all cases because this simplifies porting
to Python 3, and because there didn't seem to be any harm.

2014-09-13  Tom Tromey  <tom@tromey.com>

	* python/py-value.c (value_object_as_number): Add
	nb_inplace_divide entry, nb_inplace_floor_divide,
	nb_inplace_true_divide, nb_index.

2014-09-13  Tom Tromey  <tom@tromey.com>

	* gdb.python/py-value.exp (test_value_numeric_ops): Add test that
	uses value as an index.